requirements.txt
httpx~=0.28.1
fastapi~=0.115.11
SQLAlchemy~=2.0.38
pydantic~=2.10.6
python-dotenv~=1.0.1
pydantic-settings~=2.8.1
pytest~=8.3.5
uvicorn~=0.34.0